refactor: remove remaining old route bits
