+++ src/lib/sendErrorResponse.ts
